book type changes
